jinahub/rankers/DPRReaderRanker/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.9.0+cpu
transformers==4.9.1
jina-commons @ git+https://github.com/jina-ai/jina-commons.git@v0.0.2#egg=jina-commons
